feat: Streamline js api with cli
