+++ .vscode/settings.json
@@ -2,5 +2,6 @@
     "[javascript]": {
         "editor.formatOnSave": true
     },
-    "eslint.autoFixOnSave": true
+    "eslint.autoFixOnSave": true,
+    "editor.tabSize": 2
 }
