+++ js/__tests__/index.test.js
@@ -1,53 +1,85 @@
 /* eslint-env jest */
 
 var SentryCli = require('..');
+var helper = require('../helper');
+
+var SOURCEMAPS_OPTIONS = {
+  ignore: {
+    param: '--ignore',
+    type: 'array'
+  },
+  ignoreFile: {
+    param: '--ignore-file',
+    type: 'string'
+  },
+  rewrite: {
+    param: '--rewrite',
+    type: 'boolean'
+  },
+  sourceMapReference: {
+    param: '--no-sourcemap-reference',
+    type: 'inverted-boolean'
+  },
+  stripPrefix: {
+    param: '--strip-prefix',
+    type: 'array'
+  },
+  stripCommonPrefix: {
+    param: '--strip-common-prefix',
+    type: 'array'
+  },
+  validate: {
+    param: '--validate',
+    type: 'boolean'
+  },
+  urlPrefix: {
+    param: '--url-prefix',
+    type: 'string'
+  },
+  ext: {
+    param: '--ext',
+    type: 'string'
+  }
+};
 
 describe('SentryCli', function() {
   test('call sentry-cli --version', function() {
     expect.assertions(1);
-    var cli = new SentryCli();
-    return cli.execute(['--version']).then(function() {
+    return helper.execute(['--version']).then(function() {
       expect(true).toBe(true);
     });
   });
 
   test('call sentry-cli with wrong command', function() {
     expect.assertions(1);
-    var cli = new SentryCli();
-    return cli.execute(['fail']).catch(function(e) {
+    return helper.execute(['fail']).catch(function(e) {
       expect(e.message).toMatch('Command failed:');
     });
   });
 
   test('call prepare command add default ignore', function() {
-    var cli = new SentryCli();
-
     var command = ['releases', 'files', 'release', 'upload-sourcemaps', '/dev/null'];
 
-    expect(cli.prepareCommand(command)).toEqual([
+    expect(helper.prepareCommand(command)).toEqual([
       'releases',
       'files',
       'release',
       'upload-sourcemaps',
-      '/dev/null',
-      '--ignore',
-      'node_modules'
+      '/dev/null'
     ]);
   });
 
   test('call prepare command with array option', function() {
-    var cli = new SentryCli();
-
     var command = ['releases', 'files', 'release', 'upload-sourcemaps', '/dev/null'];
 
-    expect(cli.prepareCommand(command, { stripPrefix: ['node', 'app'] })).toEqual([
+    expect(
+      helper.prepareCommand(command, SOURCEMAPS_OPTIONS, { stripPrefix: ['node', 'app'] })
+    ).toEqual([
       'releases',
       'files',
       'release',
       'upload-sourcemaps',
       '/dev/null',
-      '--ignore',
-      'node_modules',
       '--strip-prefix',
       'node',
       '--strip-prefix',
@@ -56,82 +88,65 @@ describe('SentryCli', function() {
 
     // Should throw since it is no array
     expect(function() {
-      cli.prepareCommand(command, { stripPrefix: 'node' });
+      helper.prepareCommand(command, SOURCEMAPS_OPTIONS, { stripPrefix: 'node' });
     }).toThrow();
   });
 
   test('call prepare command with boolean option', function() {
-    var cli = new SentryCli();
-
     var command = ['releases', 'files', 'release', 'upload-sourcemaps', '/dev/null'];
 
-    expect(cli.prepareCommand(command, { sourceMapReference: false })).toEqual([
+    expect(
+      helper.prepareCommand(command, SOURCEMAPS_OPTIONS, { sourceMapReference: false })
+    ).toEqual([
       'releases',
       'files',
       'release',
       'upload-sourcemaps',
       '/dev/null',
-      '--ignore',
-      'node_modules',
       '--no-sourcemap-reference'
     ]);
 
-    expect(cli.prepareCommand(command, { sourceMapReference: true })).toEqual([
-      'releases',
-      'files',
-      'release',
-      'upload-sourcemaps',
-      '/dev/null',
-      '--ignore',
-      'node_modules'
-    ]);
+    expect(
+      helper.prepareCommand(command, SOURCEMAPS_OPTIONS, { sourceMapReference: true })
+    ).toEqual(['releases', 'files', 'release', 'upload-sourcemaps', '/dev/null']);
 
-    expect(cli.prepareCommand(command, { rewrite: true })).toEqual([
-      'releases',
-      'files',
-      'release',
-      'upload-sourcemaps',
-      '/dev/null',
-      '--ignore',
-      'node_modules',
-      '--rewrite'
-    ]);
+    expect(helper.prepareCommand(command, SOURCEMAPS_OPTIONS, { rewrite: true })).toEqual(
+      ['releases', 'files', 'release', 'upload-sourcemaps', '/dev/null', '--rewrite']
+    );
 
     expect(function() {
-      cli.prepareCommand(command, { sourceMapReference: 'node' });
+      helper.prepareCommand(command, SOURCEMAPS_OPTIONS, { sourceMapReference: 'node' });
     }).toThrow();
   });
 
   test('call prepare command with string option', function() {
-    var cli = new SentryCli();
-
     var command = ['releases', 'files', 'release', 'upload-sourcemaps', '/dev/null'];
 
-    expect(cli.prepareCommand(command, { ext: 'js' })).toEqual([
+    expect(helper.prepareCommand(command, SOURCEMAPS_OPTIONS, { ext: 'js' })).toEqual([
       'releases',
       'files',
       'release',
       'upload-sourcemaps',
       '/dev/null',
-      '--ignore',
-      'node_modules',
       '--ext',
       'js'
     ]);
 
-    expect(cli.prepareCommand(command, { urlPrefix: '~/' })).toEqual([
+    expect(
+      helper.prepareCommand(command, SOURCEMAPS_OPTIONS, { urlPrefix: '~/' })
+    ).toEqual([
       'releases',
       'files',
       'release',
       'upload-sourcemaps',
       '/dev/null',
-      '--ignore',
-      'node_modules',
       '--url-prefix',
       '~/'
     ]);
 
-    expect(cli.prepareCommand(command, { ignoreFile: '/js.ignore' })).toEqual([
+    expect(
+      helper.prepareCommand(command, SOURCEMAPS_OPTIONS, { ignoreFile: '/js.ignore' })
+    ).toEqual([
       'releases',
       'files',
       'release',
@@ -140,5 +155,13 @@ describe('SentryCli', function() {
       '--ignore-file',
       '/js.ignore'
     ]);
+  });
+
+  test('call sentry-cli releases propose-version', function() {
+    expect.assertions(1);
+    var cli = new SentryCli();
+    return cli.releases.proposeVersion().then(function(version) {
+      expect(version).toBeTruthy();
+    });
   });
 });
